+++ contracts/libraries/Math.sol
@@ -1,14 +1,14 @@
 pragma solidity 0.5.12;
 
 library Math {
-    function min(uint256 x, uint256 y) internal pure returns (uint256 z) {
+    function min(uint x, uint y) internal pure returns (uint z) {
         return x <= y ? x : y;
     }
 
-    function sqrt(uint256 x) internal pure returns (uint256 y) {
+    function sqrt(uint x) internal pure returns (uint y) {
         if (x == 0) return 0;
         else if (x <= 3) return 1;
-        uint256 z = (x + 1) / 2;
+        uint z = (x + 1) / 2;
         y = x;
         while (z < y) {
             y = z;
